+++ pyogrio/tests/test_raw_io.py
@@ -11,7 +11,7 @@ from pyogrio.errors import DriverError
 
 
 def test_read(naturalearth_lowres):
-    meta, geometry, fields = read(naturalearth_lowres)
+    meta, index, geometry, fields = read(naturalearth_lowres)
 
     assert meta["crs"] == "EPSG:4326"
     assert meta["geometry_type"] == "Polygon"
@@ -27,7 +27,9 @@ def test_read(naturalearth_lowres):
     ]
 
     assert len(fields) == 5
+    assert len(index) == len(fields[0])
     assert len(geometry) == len(fields[0])
+    assert index.dtype == np.int64
 
     # quick test that WKB is a Polygon type
     assert geometry[0][:6] == b"\x01\x06\x00\x00\x00\x03"
@@ -38,53 +40,58 @@ def test_vsi_read_layers(naturalearth_lowres_vsi):
         list_layers(naturalearth_lowres_vsi), [["naturalearth_lowres", "Polygon"]]
     )
 
-    meta, geometry, fields = read(naturalearth_lowres_vsi)
+    geometry = read(naturalearth_lowres_vsi)[2]
     assert geometry.shape == (177,)
 
 
 def test_read_no_geometry(naturalearth_lowres):
-    meta, geometry, fields = read(naturalearth_lowres, read_geometry=False)
+    geometry = read(naturalearth_lowres, read_geometry=False)[2]
 
     assert geometry is None
 
 
 def test_read_columns(naturalearth_lowres):
     # read no columns or geometry
-    meta, geometry, fields = read(naturalearth_lowres, columns=[], read_geometry=False)
+    meta, _, geometry, fields = read(
+        naturalearth_lowres, columns=[], read_geometry=False
+    )
     assert geometry is None
     assert len(fields) == 0
     array_equal(meta["fields"], np.empty(shape=(0, 4), dtype="object"))
 
     columns = ["NAME", "NAME_LONG"]
-    meta, geometry, fields = read(
+    meta, _, geometry, fields = read(
         naturalearth_lowres, columns=columns, read_geometry=False
     )
     array_equal(meta["fields"], columns)
 
     # Repeats should be dropped
     columns = ["NAME", "NAME_LONG", "NAME"]
-    meta, geometry, fields = read(
+    meta, _, geometry, fields = read(
         naturalearth_lowres, columns=columns, read_geometry=False
     )
     array_equal(meta["fields"], columns[:2])
 
 
 def test_read_skip_features(naturalearth_lowres):
-    expected_geometry, expected_fields = read(naturalearth_lowres)[1:]
-    geometry, fields = read(naturalearth_lowres, skip_features=10)[1:]
+    expected_geometry, expected_fields = read(naturalearth_lowres)[2:]
+    index, geometry, fields = read(naturalearth_lowres, skip_features=10)[1:]
 
+    assert len(index) == len(geometry)
     assert len(geometry) == len(expected_geometry) - 10
     assert len(fields[0]) == len(expected_fields[0]) - 10
 
+    assert index[0] == 10
     assert np.array_equal(geometry, expected_geometry[10:])
     # Last field has more variable data
     assert np.array_equal(fields[-1], expected_fields[-1][10:])
 
 
 def test_read_max_features(naturalearth_lowres):
-    expected_geometry, expected_fields = read(naturalearth_lowres)[1:]
-    geometry, fields = read(naturalearth_lowres, max_features=2)[1:]
+    expected_geometry, expected_fields = read(naturalearth_lowres)[2:]
+    index, geometry, fields = read(naturalearth_lowres, max_features=2)[1:]
 
+    assert len(index) == 2
     assert len(geometry) == 2
     assert len(fields[0]) == 2
 
@@ -94,13 +101,13 @@ def test_read_max_features(naturalearth_lowres):
 
 def test_read_where(naturalearth_lowres):
     # empty filter should return full set of records
-    geometry, fields = read(naturalearth_lowres, where="")[1:]
+    geometry, fields = read(naturalearth_lowres, where="")[2:]
     assert len(geometry) == 177
     assert len(fields) == 5
     assert len(fields[0]) == 177
 
     # should return singular item
-    geometry, fields = read(naturalearth_lowres, where="iso_a3 = 'CAN'")[1:]
+    geometry, fields = read(naturalearth_lowres, where="iso_a3 = 'CAN'")[2:]
     assert len(geometry) == 1
     assert len(fields) == 5
     assert len(fields[0]) == 1
@@ -109,14 +116,14 @@ def test_read_where(naturalearth_lowres):
     # should return items within range
     geometry, fields = read(
         naturalearth_lowres, where="POP_EST >= 10000000 AND POP_EST < 100000000"
-    )[1:]
+    )[2:]
     assert len(geometry) == 75
     assert min(fields[0]) >= 10000000
     assert max(fields[0]) < 100000000
 
     # should match no items
     with pytest.warns(UserWarning, match="does not have any features to read"):
-        geometry, fields = read(naturalearth_lowres, where="iso_a3 = 'INVALID'")[1:]
+        geometry, fields = read(naturalearth_lowres, where="iso_a3 = 'INVALID'")[2:]
         assert len(geometry) == 0
 
 
@@ -134,26 +141,28 @@ def test_read_bbox_invalid(naturalearth_lowres, bbox):
 def test_read_bbox(naturalearth_lowres):
     # should return no features
     with pytest.warns(UserWarning, match="does not have any features to read"):
-        geometry, fields = read(naturalearth_lowres, bbox=(0, 0, 0.00001, 0.00001))[1:]
+        geometry, fields = read(naturalearth_lowres, bbox=(0, 0, 0.00001, 0.00001))[2:]
 
     assert len(geometry) == 0
 
-    geometry, fields = read(naturalearth_lowres, bbox=(-140, 20, -100, 40))[1:]
+    geometry, fields = read(naturalearth_lowres, bbox=(-140, 20, -100, 40))[2:]
 
     assert len(geometry) == 2
     assert np.array_equal(fields[3], ["USA", "MEX"])
 
 
 def test_read_fids(naturalearth_lowres):
-    expected_geometry, expected_fields = read(naturalearth_lowres)[1:]
+    expected_fids, expected_geometry, expected_fields = read(naturalearth_lowres)[1:]
     subset = [0, 10, 5]
-    
-    for fids in [subset, np.array(subset)]:
-        geometry, fields = read(naturalearth_lowres, fids=subset)[1:]
 
+    for fids in [subset, np.array(subset)]:
+        index, geometry, fields = read(naturalearth_lowres, fids=subset)[1:]
+
+        assert len(fids) == 3
         assert len(geometry) == 3
         assert len(fields[0]) == 3
 
+        assert np.array_equal(fids, expected_fids[subset])
         assert np.array_equal(geometry, expected_geometry[subset])
         assert np.array_equal(fields[-1], expected_fields[-1][subset])
 
@@ -181,7 +190,7 @@ def test_read_fids_unsupported_keywords(naturalearth_lowres):
 
 
 def test_write(tmpdir, naturalearth_lowres):
-    meta, geometry, field_data = read(naturalearth_lowres)
+    meta, index, geometry, field_data = read(naturalearth_lowres)
 
     filename = os.path.join(str(tmpdir), "test.shp")
     write(filename, geometry, field_data, **meta)
@@ -192,7 +201,7 @@ def test_write(tmpdir, naturalearth_lowres):
 
 
 def test_write_gpkg(tmpdir, naturalearth_lowres):
-    meta, geometry, field_data = read(naturalearth_lowres)
+    meta, index, geometry, field_data = read(naturalearth_lowres)
 
     filename = os.path.join(str(tmpdir), "test.gpkg")
     write(filename, geometry, field_data, driver="GPKG", **meta)
@@ -201,7 +210,7 @@ def test_write_gpkg(tmpdir, naturalearth_lowres):
 
 
 def test_write_geojson(tmpdir, naturalearth_lowres):
-    meta, geometry, field_data = read(naturalearth_lowres)
+    meta, index, geometry, field_data = read(naturalearth_lowres)
 
     filename = os.path.join(str(tmpdir), "test.json")
     write(filename, geometry, field_data, driver="GeoJSON", **meta)
@@ -229,7 +238,7 @@ def test_write_geojson(tmpdir, naturalearth_lowres):
 )
 def test_write_supported(tmpdir, naturalearth_lowres, driver):
     """Test drivers not specifically tested above"""
-    meta, geometry, field_data = read(naturalearth_lowres, columns=["iso_a3"])
+    meta, index, geometry, field_data = read(naturalearth_lowres, columns=["iso_a3"])
 
     # note: naturalearth_lowres contains mixed polygons / multipolygons, which
     # are not supported in mixed form for all drivers.  To get around this here
@@ -242,14 +251,14 @@ def test_write_supported(tmpdir, naturalearth_lowres, driver):
         geometry[:1],
         field_data=[f[:1] for f in field_data],
         driver=driver,
-        **meta
+        **meta,
     )
 
     assert filename.exists()
 
 
 def test_write_unsupported(tmpdir, naturalearth_lowres):
-    meta, geometry, field_data = read(naturalearth_lowres)
+    meta, index, geometry, field_data = read(naturalearth_lowres)
 
     filename = os.path.join(str(tmpdir), "test.fgdb")